backend/requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
Flask==2.2.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
PyJWT==2.4.0
MarkupSafe==2.1.1
pycparser==2.21
requests==2.28.1
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
